+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.0](#030)
   - [0.2.0](#020)
   - [0.1.0](#010)
 
 ---
+
+## 0.3.0
+
+Released on 29/09/2024
+
+- Added `KubeMultiPodFs` to operate on multiple pod and containers at the same time. See docs for details.
+- **BREAKING ‼️** Renamed `KubeFs` to `KubeContainerFs`.
 
 ## 0.2.0